+++ src/collective/cart/shipping/interfaces.py
@@ -1,40 +1,20 @@
-from collective.cart.shipping import _
-from zope import schema
+from collective.cart.shipping.schema import OrderShippingMethodSchema
+from collective.cart.shipping.schema import ShippingMethodContainerSchema
+from collective.cart.shipping.schema import ShippingMethodSchema
+from plone.dexterity.interfaces import IDexterityContainer
 from zope.interface import Attribute
-from zope.interface import Interface
 
 
-class IShippingMethodContainer(Interface):
-    """Interface for ShippingMethodContainer"""
+class IShippingMethodContainer(ShippingMethodContainerSchema, IDexterityContainer):
+    """Interface for content type: collective.cart.shipping.ShippingMethodContainer"""
 
 
-class IShippingMethod(Interface):
-
-    to_country = schema.Choice(
-        title=_(u"Country To"),
-        required=False,
-        description=_(u"Select countries to which this shipping method is applied."),
-        vocabulary=_(u"Countries"))
-
-    min_delivery_days = schema.Int(
-        title=_(u"Minimum Delivery Days"),
-        required=True)
-
-    max_delivery_days = schema.Int(
-        title=_(u"Maximum Delivery Days"),
-        required=True)
-
-    weight_dimension_rate = schema.Float(
-        title=_(u"Weight Dimension Rate"),
-        description=_(u"1 m3 = ??? kg"),
-        required=False)
-
-    vat = schema.Choice(
-        title=_(u'VAT'),
-        vocabulary=u'collective.behavior.vat.rates')
+class IShippingMethod(ShippingMethodSchema):
+    """Interface for content type: ShippingMethod"""
 
 
-class ICartShippingMethod(Interface):
+class IOrderShippingMethod(OrderShippingMethodSchema, IDexterityContainer):
+    """Interface for content type: collective.cart.shipping.OrderShippingMethod"""
 
     title = Attribute(u'Title')
     min_delivery_days = Attribute(u"Minimum Delivery Days")